requirements.txt
timm>=0.6.1
numpy>=1.23.1
opencv_python>=4.6.0.66
loguru>=0.6.0
scikit-image>=0.19.3
tqdm>=4.64.1
Pillow>=9.2.0
thop>=0.1.1.post2209072238
ninja>=1.10.2.4
tabulate>=0.8.10
tensorboard>=2.10.1
attrdict>=2.0.1
pytest-cov>=4.0.0
fast_coco_eval>=1.0
albumentations>=1.1.0
nltk>=3.7
lmdb>=1.3.0
natsort>=8.2.0
torchinfo>=1.7.0
pandas>=1.5.0
optuna>=3.0.3

onnx>=1.12.0
onnxruntime>=1.11.1
onnx-simplifier>=0.4.0
